Move agent tests into it's own package
Workers depending on cmd package is stupid.
