+++ agent/engine/config.go
@@ -1,3 +1,6 @@
+// Copyright 2023 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
 package engine
 
 import (
