+++ cmd/containeragent/unit/manifolds_test.go
@@ -9,8 +9,8 @@ import (
 	gc "gopkg.in/check.v1"
 
 	"github.com/juju/juju/agent"
+	"github.com/juju/juju/agent/agenttest"
 	"github.com/juju/juju/cmd/containeragent/unit"
-	"github.com/juju/juju/cmd/jujud/agent/agenttest"
 	"github.com/juju/juju/testing"
 )
