+++ cmd/jujud/agent/modeloperator/manifolds_test.go
@@ -8,7 +8,7 @@ import (
 	gc "gopkg.in/check.v1"
 
 	"github.com/juju/juju/agent"
-	"github.com/juju/juju/cmd/jujud/agent/agenttest"
+	"github.com/juju/juju/agent/agenttest"
 	"github.com/juju/juju/cmd/jujud/agent/modeloperator"
 	"github.com/juju/juju/testing"
 )
